pyproject.toml
[project]
name = "gack"
version = "0.1.0"
description = "Privacy preserving home monitoring"
dependencies = [
    "ultralytics>=8.3.169",
    "opencv-python>=4.12.0.88",
    "ffmpeg-python>=0.2.0",
    "python-dotenv>=1.1.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "requests>=2.32.4",
    "numpy<=2.2.6",
    "aiosqlite>=0.20.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/gack"]

[project.scripts]
gack = "gack.__main__:main"
gack-web = "gack.web_server:main"
